refactor: move server queryables
